Рефакторинг. Изменена обработка символов экранирования.
