+++ cs/Markdown/Extensions/StringExtensions.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Markdown.Extensions
+{
+    public static class StringExtensions
+    {
+        public static string RemoveEscapeSymbols(this string text) => text.Replace(@"\", "");
+
+        public static List<int> AllIndexesOf(this string str, string value)
+        {
+            if (string.IsNullOrEmpty(value))
+                throw new ArgumentException("the string to find may not be empty", nameof(value));
+
+            var indexes = new List<int>();
+            for (var index = 0; ; index += value.Length)
+            {
+                index = str.IndexOf(value, index, StringComparison.Ordinal);
+                if (index == -1)
+                    return indexes;
+                indexes.Add(index);
+            }
+        }
+    }
+}
